delete RecipesController and RecipesModel files
