+++ app/Model/User.php
@@ -268,15 +268,11 @@ class User
         // session_start();
 
         $userId = $_SESSION['id'];
-
         $sql = "SELECT * FROM recipes WHERE name LIKE :searchValue AND (creator = :userId OR creator = 42)";
-
         $this->db->query($sql);
         $this->db->bind(':searchValue', "%$searchValue%");
         $this->db->bind(':userId', $userId);
-
         $results = $this->db->resultSet();
-
         if ($this->db->rowCount() > 0) {
             return $results;
         } else {
@@ -299,13 +295,10 @@ class User
         $this->db->query($sql);
         $this->db->bind(':userId', $userId);
         $rows = $this->db->resultSet();
-
+       $data[]= 0;
         if ($this->db->rowCount() >= 0) {
             foreach ($rows as $row) {
                 $data[] = $row;
-            }
-            if ($data == null) {
-                $data = 0;
             }
             $_SESSION['notifications'] = $data;
             return $this->db->rowCount();
@@ -313,4 +306,50 @@ class User
             return false;
         }
     }
+    /**
+     * get all recipes
+     *
+     * @return array|bool
+     */
+    function getRecipesList()
+    {
+        $sql = "SELECT * FROM recipes";
+        $this->db->query($sql);
+        $row = $this->db->resultSet();
+        if ($this->db->rowCount() > 0) {
+            return $row;
+        } else {
+            return false;
+        }
+    }
+    /**
+     * add Recipe
+     * 
+     * Add recipe from the parameters, in the database
+     *
+     * @param  mixed $donnees
+     * @return bool
+     */
+    function addRecipe($donnees)
+    {
+
+        $sql = "INSERT INTO  recipes(name,calories,image_url) VALUES (:name, :calories, :image_url )";
+        $this->db->query($sql);
+        $this->db->bind(':name', $donnees['name']);
+        $this->db->bind(':calories', $donnees['calories']);
+        $this->db->bind(':image_url', $donnees['image_url']);
+
+        try {
+            if ($this->db->execute()) {
+                return true;
+            } else {
+                return false;
+            }
+        } catch (\PDOException $e) {
+            // Handle exception
+            echo "Database error: " . $e->getMessage();
+            return false;
+        }
+    }
 }
+
